layer: Combine 2 Instruction Classes
