+++ layers/gpuav/spirv/descriptor_class_texel_buffer_pass.cpp
@@ -200,7 +200,7 @@ bool DescriptorClassTexelBufferPass::Instrument() {
                 // Add any debug information to pass into the function call
                 InjectionData injection_data;
                 injection_data.stage_info_id = GetStageInfo(*function, block_it, inst_it);
-                const uint32_t inst_position = target_instruction_->position_index_;
+                const uint32_t inst_position = target_instruction_->GetPositionIndex();
                 auto inst_position_constant = module_.type_manager_.CreateConstantUInt32(inst_position);
                 injection_data.inst_position_id = inst_position_constant.Id();
