+++ layers/gpuav/spirv/inject_conditional_function_pass.cpp
@@ -154,7 +154,7 @@ bool InjectConditionalFunctionPass::Instrument() {
                 // Add any debug information to pass into the function call
                 InjectionData injection_data;
                 injection_data.stage_info_id = GetStageInfo(*function, block_it, inst_it);
-                const uint32_t inst_position = target_instruction_->position_index_;
+                const uint32_t inst_position = target_instruction_->GetPositionIndex();
                 auto inst_position_constant = module_.type_manager_.CreateConstantUInt32(inst_position);
                 injection_data.inst_position_id = inst_position_constant.Id();
