+++ layers/state_tracker/shader_instruction.h
@@ -23,6 +23,11 @@
 #include "containers/custom_containers.h"
 #include <spirv/unified1/spirv.hpp>
 
+struct OperandInfo;
+
+// Need to use until we have native std::span in c++20
+using spirv_iterator = vvl::enumeration<const uint32_t, const uint32_t*>::iterator;
+
 namespace spirv {
 
 // Holds information about a single SPIR-V instruction
@@ -39,6 +44,8 @@ class Instruction {
 
     // The word used to define the Instruction
     uint32_t Word(uint32_t index) const { return words_[index]; }
+    // Skips pass any optional Result or Result Type word
+    uint32_t Operand(uint32_t index) const { return words_[operand_index_ + index]; }
 
     uint32_t Length() const { return words_[0] >> 16; }
 
@@ -64,8 +71,9 @@ class Instruction {
     uint32_t GetBitWidth() const;
     uint32_t GetByteWidth() const { return (GetBitWidth() + 31) / 32; }
     spv::BuiltIn GetBuiltIn() const;
-
+    uint32_t GetPositionIndex() const { return position_index_; }
     bool IsArray() const;
+    bool IsAccessChain() const;
     // Helpers for OpTypeImage
     spv::Dim FindImageDim() const;
     bool IsImageArray() const;
@@ -76,6 +84,20 @@ class Instruction {
 
     bool operator==(Instruction const& other) const { return words_ == other.words_; }
     bool operator!=(Instruction const& other) const { return words_ != other.words_; }
+
+    // The following is only used for GPU-AV where we need to possibly update an Instruction
+    Instruction(spirv_iterator it, uint32_t position);
+    // Assumes caller will fill remaining words
+    Instruction(uint32_t length, spv::Op opcode);
+    void Fill(const std::vector<uint32_t>& words);
+    void UpdateWord(uint32_t index, uint32_t data);
+    void ToBinary(std::vector<uint32_t>& out);
+    // Increments Length() as well
+    void AppendWord(uint32_t word);
+    void ReplaceResultId(uint32_t new_result_id);
+    // searchs all operands to replace ID if found
+    void ReplaceOperandId(uint32_t old_word, uint32_t new_word);
+    void ReplaceLinkedId(vvl::unordered_map<uint32_t, uint32_t>& id_swap_map);
 
   private:
     void SetResultTypeIndex();
@@ -92,6 +114,11 @@ class Instruction {
     small_vector<uint32_t, word_vector_length, uint32_t> words_;
     uint32_t result_id_index_ = 0;
     uint32_t type_id_index_ = 0;
+    uint32_t operand_index_ = 1;
+
+    // used to find original position of instruction in shader, pre-instrumented
+    const uint32_t position_index_;
+    const OperandInfo& operand_info_;
 
 #ifndef NDEBUG
     // Helping values to make debugging what is happening in a instruction easier
